Move handle types to public headers
