+++ include/ccf/kv/abstract_handle.h
@@ -1,0 +1,12 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the Apache 2.0 License.
+#pragma once
+
+namespace kv
+{
+  class AbstractHandle
+  {
+  public:
+    virtual ~AbstractHandle() = default;
+  };
+}
