+++ include/ccf/kv/unit.h
@@ -2,7 +2,7 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
-#include "kv/serialised_entry.h"
+#include "ccf/kv/serialised_entry.h"
 
 namespace kv::serialisers
 {
