+++ src/kv/apply_changes.h
@@ -5,7 +5,6 @@
 #include "ccf/tx.h"
 #include "kv/untyped_change_set.h"
 #include "kv_types.h"
-#include "untyped_map_handle.h"
 
 #include <functional>
 #include <map>
