+++ src/kv/map_handle.h
@@ -2,7 +2,7 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
-#include "kv/untyped_map_handle.h"
+#include "ccf/kv/untyped_map_handle.h"
 
 namespace kv
 {
