+++ src/kv/serialise_entry_blit.h
@@ -4,7 +4,7 @@
 
 #include "ccf/ds/nonstd.h"
 #include "ccf/entity_id.h"
-#include "serialised_entry.h"
+#include "ccf/kv/serialised_entry.h"
 
 namespace kv::serialisers
 {
