+++ src/kv/serialise_entry_json.h
@@ -2,7 +2,7 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
-#include "serialised_entry.h"
+#include "ccf/kv/serialised_entry.h"
 
 #include <nlohmann/json.hpp>
