+++ src/kv/untyped_map_handle.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
 
-#include "kv/untyped_map_handle.h"
+#include "ccf/kv/untyped_map_handle.h"
 
 #include "kv/untyped_change_set.h"
